tests/wpt/tests/fedcm/fedcm-domainhint-matches-account.https.html

<!DOCTYPE html>
<title>Federated Credential Management API domain hint matches an account.</title>
<link rel="help" href="https://fedidcg.github.io/FedCM">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<body>

<script type="module">
import {
  fedcm_test,
  request_options_with_domain_hint,
  fedcm_get_and_select_first_account
} from './support/fedcm-helper.sub.js';

fedcm_test(async t => {
  const options = request_options_with_domain_hint('manifest.py',
    'idp.example');
  const cred = await fedcm_get_and_select_first_account(t, options);
  assert_equals(cred.token, 'token');
}, "Domain hint matches an account.");
  </script>
